# src/main/resources/openapi/station-search-api.yml
openapi: 3.0.3
info:
  title: Station Search API
  version: 1.0.0

paths:
  /stations:
    get:
      tags:
        - Stations
      operationId: searchStations
      summary: Search stations by radius and filters
      description: |
        Searches stations using one of two mutually exclusive modes.

        VIEWPORT mode:
        - Requires north, south, east and west.
        - Must not include lat, lng or radiusMeters.
        - DISTANCE sorting is not supported because there is no reference point.

        RADIUS mode:
        - Requires lat, lng and radiusMeters.
        - Must not include north, south, east or west.
        - Supports PRICE and DISTANCE sorting.

        Supplying incomplete parameters or mixing parameters from both modes
        returns HTTP 400.

      parameters:
        - name: page
          in: query
          required: false
          description: Zero-based page number
          schema:
            type: integer
            format: int32
            minimum: 0
            default: 0

        - name: size
          in: query
          required: false
          description: Number of stations returned per page
          schema:
            type: integer
            format: int32
            minimum: 1
            maximum: 100
            default: 50

        - name: searchMode
          in: query
          required: true
          description: Defines how the geographical search area is specified.
          schema:
            $ref: "#/components/schemas/StationSearchMode"

        # Radius filters
        - name: lat
          in: query
          required: false
          schema:
            type: number
            format: double
            minimum: -90
            maximum: 90
        - name: lng
          in: query
          required: false
          schema:
            type: number
            format: double
            minimum: -180
            maximum: 180
        - name: radiusMeters
          in: query
          required: false
          schema:
            type: integer
            format: int32
            minimum: 1
            maximum: 50000

        # Viewport parameters
        - name: north
          in: query
          required: false
          description: |
              Northern latitude boundary of the visible map.
              Required when searchMode is VIEWPORT.
          schema:
            type: number
            format: double
            minimum: -90
            maximum: 90
          example: 36.7601
        - name: south
          in: query
          required: false
          description: |
            Southern latitude boundary of the visible map.
            Required when searchMode is VIEWPORT.
            Must be lower than north.
          schema:
            type: number
            format: double
            minimum: -90
            maximum: 90
          example: 36.6802
        - name: east
          in: query
          required: false
          description: |
            Eastern longitude boundary of the visible map.
            Required when searchMode is VIEWPORT.
          schema:
            type: number
            format: double
            minimum: -180
            maximum: 180
          example: -4.3501
        - name: west
          in: query
          required: false
          description: |
            Western longitude boundary of the visible map.
            Required when searchMode is VIEWPORT.
          schema:
            type: number
            format: double
            minimum: -180
            maximum: 180
          example: -4.4907

        # Common filters
        - name: productType
          in: query
          required: false
          schema:
            $ref: "#/components/schemas/ProductType"
        - name: sortBy
          in: query
          required: false
          schema:
            $ref: "#/components/schemas/StationSearchSortBy"

      responses:
        "200":
          description: Stations found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StationSearchResponse"

  /locations/search:
    get:
      tags:
        - Locations
      summary: Search locations
      description: >
        Returns location suggestions matching a locality name or postal code.
        The returned coordinates can be used to center the map.
      operationId: searchLocations

      parameters:
        - name: query
          in: query
          required: true
          description: Locality name or postal code to search.
          schema:
            type: string
            minLength: 2
          example: Malaga
        - name: countryCode
          in: query
          required: false
          description: ISO 3166-1 alpha-2 country code.
          schema:
            type: string
            pattern: '^[A-Z]{2}$'
            default: ES
          example: ES
        - name: limit
          in: query
          required: false
          description: Maximum number of suggestions to return.
          schema:
            type: integer
            format: int32
            minimum: 1
            maximum: 20
            default: 10
          example: 10

      responses:
        '200':
          description: Location suggestions found.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/LocationSuggestion'

  /statistics/fuel-prices/current:
    get:
      tags: [Fuel Price Statistics]
      operationId: getCurrentFuelPriceStatistics
      parameters:
        - {name: countryCode, in: query, required: true, schema: {type: string, minLength: 2, maxLength: 2}}
        - {name: productType, in: query, required: true, schema: {$ref: '#/components/schemas/ProductType'}}
        - {name: level, in: query, required: false, schema: {$ref: '#/components/schemas/GeographicLevel', default: NATIONAL}}
        - {name: area, in: query, required: false, schema: {type: string}}
        - {name: province, in: query, required: false, description: Required for municipality scope, schema: {type: string}}
      responses:
        '200':
          description: Current aggregate
          content:
            application/json:
              schema: {$ref: '#/components/schemas/CurrentPriceStatisticsResponse'}

  /statistics/fuel-prices/history:
    get:
      tags: [Fuel Price Statistics]
      operationId: getHistoricalFuelPriceStatistics
      parameters:
        - {name: countryCode, in: query, required: true, schema: {type: string}}
        - {name: productType, in: query, required: true, schema: {$ref: '#/components/schemas/ProductType'}}
        - {name: level, in: query, required: false, schema: {$ref: '#/components/schemas/GeographicLevel', default: NATIONAL}}
        - {name: area, in: query, required: false, schema: {type: string}}
        - {name: province, in: query, required: false, schema: {type: string}}
        - {name: from, in: query, required: true, schema: {type: string, format: date}}
        - {name: to, in: query, required: true, schema: {type: string, format: date}}
      responses:
        '200':
          description: Daily historical aggregates
          content:
            application/json:
              schema:
                type: array
                items: {$ref: '#/components/schemas/HistoricalPricePointResponse'}

  /statistics/fuel-prices/provinces:
    get:
      tags: [Fuel Price Statistics]
      operationId: getProvinceFuelPriceStatistics
      parameters:
        - {name: countryCode, in: query, required: true, schema: {type: string}}
        - {name: productType, in: query, required: true, schema: {$ref: '#/components/schemas/ProductType'}}
      responses:
        '200':
          description: Ranked province aggregates
          content:
            application/json:
              schema:
                type: array
                items: {$ref: '#/components/schemas/RankedAreaStatisticsResponse'}

  /statistics/fuel-prices/summary:
    get:
      tags:
        - Fuel Price Statistics
      operationId: getFuelPriceSummary
      summary: Get fuel price summary
      description: >
        Returns aggregated fuel price statistics for a country and product,
        including average price, variation, minimum and maximum prices and number of stations analysed.

      parameters:
        - name: countryCode
          in: query
          required: true
          schema:
            type: string
            minLength: 2
            maxLength: 2
            example: ES
        - name: productType
          in: query
          required: true
          schema:
            $ref: '#/components/schemas/ProductType'
        - name: days
          in: query
          required: false
          description: Number of previous days used to calculate the statistics.
          schema:
            type: integer
            minimum: 1
            maximum: 365
            default: 7
          example: 7
      responses:
        '200':
          description: Fuel price summary
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FuelPriceSummaryResponse'

  /statistics/fuel-prices/savings:
    get:
      tags: [Fuel Price Statistics]
      operationId: getFuelSaving
      parameters:
        - {name: stationId, in: query, required: true, schema: {type: string, format: uuid}}
        - {name: productType, in: query, required: true, schema: {$ref: '#/components/schemas/ProductType'}}
        - {name: referencePrice, in: query, required: true, schema: {type: number}}
        - {name: tankLiters, in: query, required: false, schema: {type: number, default: 55}}
      responses:
        '200':
          description: Theoretical saving against the reference price
          content:
            application/json:
              schema: {$ref: '#/components/schemas/FuelSavingResponse'}

components:

  schemas:

    FuelPriceSummaryResponse:
      type: object
      required:
        - countryCode
        - productType
        - averagePrice
        - stationsAnalyzed
        - updatedAt
      properties:
        countryCode:
          type: string
          example: ES
        productType:
          $ref: '#/components/schemas/ProductType'
        averagePrice:
          type: number
          format: double
          example: 1.389
        previousAveragePrice:
          type: number
          format: double
          example: 1.409
        variation:
          type: number
          format: double
          description: Difference against the beginning of the selected period.
          example: -0.020
        variationPercentage:
          type: number
          format: double
          example: -1.42
        minimumPrice:
          type: number
          format: double
          example: 1.358
        maximumPrice:
          type: number
          format: double
          example: 1.439
        stationsAnalyzed:
          type: integer
          format: int64
          example: 12458
        estimatedSaving:
          $ref: '#/components/schemas/EstimatedSaving'
        updatedAt:
          type: string
          format: date-time
          example: '2026-08-31T12:30:00Z'

    FuelSavingResponse:
      type: object
      properties:
        stationPrice: {type: number}
        referencePrice: {type: number}
        savingPerLiter: {type: number}
        tankLiters: {type: number, default: 55}
        savingPerTank: {type: number}

    GeographicLevel:
      type: string
      enum: [NATIONAL, PROVINCE, MUNICIPALITY]

    GeographicScopeResponse:
      type: object
      properties:
        level: {$ref: '#/components/schemas/GeographicLevel'}
        name: {type: string, nullable: true}
        province: {type: string, nullable: true}

    StationPricePointResponse:
      type: object
      properties:
        stationId: {type: string, format: uuid}
        brand: {type: string}
        price: {type: number}
        latitude: {type: number}
        longitude: {type: number}

    CurrentPriceStatisticsResponse:
      type: object
      properties:
        countryCode: {type: string}
        productType: {$ref: '#/components/schemas/ProductType'}
        scope: {$ref: '#/components/schemas/GeographicScopeResponse'}
        averagePrice: {type: number}
        minimumPrice: {type: number}
        maximumPrice: {type: number}
        stationCount: {type: integer, format: int64}
        cheapestStation: {$ref: '#/components/schemas/StationPricePointResponse'}
        mostExpensiveStation: {$ref: '#/components/schemas/StationPricePointResponse'}
        nationalAverageDifference: {type: number, nullable: true}
        provincialAverageDifference: {type: number, nullable: true}
        updatedAt: {type: string, format: date-time}

    HistoricalPricePointResponse:
      type: object
      properties:
        date: {type: string, format: date}
        averagePrice: {type: number}
        minimumPrice: {type: number}
        maximumPrice: {type: number}
        stationCount: {type: integer, format: int64}
        periodAveragePrice: {type: number}
        periodMinimumPrice: {type: number}
        periodMaximumPrice: {type: number}
        changeFromPreviousDay: {type: number, nullable: true}
        changeFromSevenDaysAgo: {type: number, nullable: true}
        changeFromThirtyDaysAgo: {type: number, nullable: true}
        percentageFromPreviousDay: {type: number, nullable: true}
        percentageFromSevenDaysAgo: {type: number, nullable: true}
        percentageFromThirtyDaysAgo: {type: number, nullable: true}

    RankedAreaStatisticsResponse:
      type: object
      properties:
        area: {type: string}
        averagePrice: {type: number}
        minimumPrice: {type: number}
        maximumPrice: {type: number}
        stationCount: {type: integer, format: int64}
        cheapestStation: {$ref: '#/components/schemas/StationPricePointResponse'}
        nationalAverageDifference: {type: number}
        cheapestRank: {type: integer}
        mostExpensiveRank: {type: integer}

    EstimatedSaving:
      type: object
      properties:
        pricePerLiter:
          type: number
          format: double
          example: 0.136
        tankSizeLiters:
          type: number
          format: double
          example: 55
        amount:
          type: number
          format: double
          example: 7.48

    LocationSuggestion:
      type: object
      required:
        - type
        - primaryText
        - secondaryText
        - countryCode
        - latitude
        - longitude
      properties:

        type:
          $ref: '#/components/schemas/LocationSuggestionType'

        primaryText:
          type: string
          description: Main text displayed in the autocomplete result.
          example: Málaga

        secondaryText:
          type: string
          description: Administrative context displayed below the main text.
          example: Málaga, Andalucía

        countryCode:
          type: string
          description: ISO 3166-1 alpha-2 country code.
          pattern: '^[A-Z]{2}$'
          example: ES

        postalCode:
          type: string
          nullable: true
          description: Postal code when the suggestion represents a postal code.
          example: '29001'

        latitude:
          type: number
          format: double
          minimum: -90
          maximum: 90
          example: 36.7213

        longitude:
          type: number
          format: double
          minimum: -180
          maximum: 180
          example: -4.4214

    LocationSuggestionType:
      type: string
      description: Type of location suggestion.
      enum:
        - LOCALITY
        - POSTAL_CODE

    StationSearchMode:
      type: string
      description: Geographical strategy used to search stations.
      enum:
        - VIEWPORT
        - RADIUS
    ProductType:
      type: string
      enum:
        - DIESEL_A
        - DIESEL_B
        - DIESEL_PREMIUM
        - RENEWABLE_DIESEL

        - GASOLINE_95_E5
        - GASOLINE_95_E10
        - GASOLINE_95_E25
        - GASOLINE_95_E5_PREMIUM
        - GASOLINE_95_E85
        - GASOLINE_98_E5
        - GASOLINE_98_E10
        - RENEWABLE_GASOLINE

        - BIODIESEL
        - BIOETHANOL

        - CNG
        - LNG
        - BIOMETHANE_CNG
        - BIOMETHANE_LNG
        - LPG
        - HYDROGEN

        - ADBLUE
        - AMMONIA
        - METHANOL

    StationSearchSortBy:
      type: string
      enum:
        - DISTANCE
        - PRICE

    StationSearchResponse:
      type: object
      required:
        - stations
        - page
      properties:
        stations:
          type: array
          items:
            $ref: "#/components/schemas/StationSearchItemResponse"
        page:
          $ref: '#/components/schemas/PageMetadata'

    PageMetadata:
      type: object
      required:
        - number
        - size
        - numberOfElements
        - hasNext
        - hasPrevious
      properties:
        number:
          type: integer
          format: int32
          example: 0
        size:
          type: integer
          format: int32
          example: 50
        numberOfElements:
          type: integer
          format: int32
          example: 50
        hasNext:
          type: boolean
          example: true
        hasPrevious:
          type: boolean
          example: false

    StationSearchItemResponse:
      type: object
      required:
        - id
        - brand
        - latitude
        - longitude
        - distanceMeters
        - productPrices
      properties:
        id:
          type: string
          format: uuid
        brand:
          type: string
        normalizedBrand:
          type: string
        address:
          $ref: "#/components/schemas/AddressResponse"
        latitude:
          type: number
          format: double
        longitude:
          type: number
          format: double
        distanceMeters:
          type: number
          format: double
        openingPeriods:
          type: array
          items:
            $ref: "#/components/schemas/OpeningPeriodResponse"
        productPrices:
          type: array
          items:
            $ref: "#/components/schemas/ProductPriceResponse"

    OpeningPeriodResponse:
      type: object
      required:
        - days
        - open
        - close
      properties:
        days:
          type: array
          items:
            type: string
            enum:
              - MON
              - TUE
              - WED
              - THU
              - FRI
              - SAT
              - SUN
        open:
          type: string
          format: time
          example: "08:00:00"
        close:
          type: string
          format: time
          example: "22:00:00"

    AddressResponse:
      type: object
      properties:
        street:
          type: string
        postalCode:
          type: string
        locality:
          type: string
        municipality:
          type: string
        province:
          type: string

    ProductPriceResponse:
      type: object
      required:
        - productType
        - price
      properties:
        productType:
          $ref: "#/components/schemas/ProductType"
        price:
          type: number
          format: double